+++ sdk/identity/identity/src/credentials/authorizationCodeCredentialOptions.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { TokenCredentialOptions } from "../client/identityClient";
-
-/**
- * Options for the {@link AuthorizationCodeCredential}
- */
-export interface AuthorizationCodeCredentialOptions extends TokenCredentialOptions {}