+++ sdk/identity/identity/src/msal/nodeFlows/msalOpenBrowser.ts
@@ -3,7 +3,7 @@
 
 import * as msalNode from "@azure/msal-node";
 
-import { AccessToken, GetTokenOptions } from "@azure/core-auth";
+import { AccessToken } from "@azure/core-auth";
 
 import { Socket } from "net";
 import http from "http";
@@ -14,6 +14,7 @@ import { credentialLogger, formatError, formatSuccess } from "../../util/logging
 import { MsalNodeOptions, MsalNode } from "./nodeCommon";
 import { msalToPublic } from "../utils";
 import { CredentialUnavailableError } from "../../client/errors";
+import { CredentialFlowGetTokenOptions } from "../credentials";
 
 /**
  * Options that can be passed to configure MSAL to handle authentication through opening a browser window.
@@ -60,7 +61,10 @@ export class MsalOpenBrowser extends MsalNode {
     return this.publicApp!.acquireTokenByCode(request);
   }
 
-  protected doGetToken(scopes: string[], options?: GetTokenOptions): Promise<AccessToken> {
+  protected doGetToken(
+    scopes: string[],
+    options?: CredentialFlowGetTokenOptions
+  ): Promise<AccessToken> {
     return new Promise<AccessToken>((resolve, reject) => {
       const socketToDestroy: Socket[] = [];
 
@@ -87,7 +91,8 @@ export class MsalOpenBrowser extends MsalNode {
         const tokenRequest: msalNode.AuthorizationCodeRequest = {
           code: url.searchParams.get("code")!,
           redirectUri: this.redirectUri,
-          scopes: scopes
+          scopes: scopes,
+          authority: options?.authority
         };
 
         this.acquireTokenByCode(tokenRequest)
@@ -167,7 +172,7 @@ export class MsalOpenBrowser extends MsalNode {
       app.on("connection", (socket) => socketToDestroy.push(socket));
 
       app.on("listening", () => {
-        const openPromise = this.openAuthCodeUrl(scopes);
+        const openPromise = this.openAuthCodeUrl(scopes, options);
 
         const abortSignal = options?.abortSignal;
         if (abortSignal) {
@@ -185,10 +190,14 @@ export class MsalOpenBrowser extends MsalNode {
     });
   }
 
-  private async openAuthCodeUrl(scopeArray: string[]): Promise<void> {
+  private async openAuthCodeUrl(
+    scopeArray: string[],
+    options?: CredentialFlowGetTokenOptions
+  ): Promise<void> {
     const authCodeUrlParameters: msalNode.AuthorizationUrlRequest = {
       scopes: scopeArray,
-      redirectUri: this.redirectUri
+      redirectUri: this.redirectUri,
+      authority: options?.authority
     };
 
     const response = await this.publicApp!.getAuthCodeUrl(authCodeUrlParameters);