+++ sdk/identity/identity/test/internal/utils.spec.ts
@@ -9,8 +9,7 @@ describe("Identity utilities", function() {
     it("throws if multi-tenant authentication is disallowed, and the tenants don't match", async function() {
       let error: Error | undefined;
       try {
-        processMultiTenantRequest("credential-options-tenant-id", {
-          allowMultiTenantAuthentication: false,
+        processMultiTenantRequest("credential-options-tenant-id", false, {
           tenantId: "get-token-options-tenant-id"
         });
       } catch (e) {
@@ -26,8 +25,7 @@ describe("Identity utilities", function() {
     it("throws if multi-tenant authentication is undefined, and the tenants don't match", async function() {
       let error: Error | undefined;
       try {
-        processMultiTenantRequest("credential-options-tenant-id", {
-          allowMultiTenantAuthentication: undefined,
+        processMultiTenantRequest("credential-options-tenant-id", undefined, {
           tenantId: "get-token-options-tenant-id"
         });
       } catch (e) {
@@ -42,15 +40,13 @@ describe("Identity utilities", function() {
 
     it("If allowMultiTenantAuthentication is disallowed, it shouldn't throw if the tenant received is the same as the tenant we already had, that same tenant should be returned", async function() {
       assert.equal(
-        processMultiTenantRequest("same-tenant", {
-          allowMultiTenantAuthentication: false,
+        processMultiTenantRequest("same-tenant", false, {
           tenantId: "same-tenant"
         }),
         "same-tenant"
       );
       assert.equal(
-        processMultiTenantRequest("same-tenant", {
-          allowMultiTenantAuthentication: undefined,
+        processMultiTenantRequest("same-tenant", undefined, {
           tenantId: "same-tenant"
         }),
         "same-tenant"
@@ -59,8 +55,7 @@ describe("Identity utilities", function() {
 
     it("If we had a tenant and the options have another tenant, we pick the tenant from the options", async function() {
       assert.equal(
-        processMultiTenantRequest("credential-options-tenant-id", {
-          allowMultiTenantAuthentication: true,
+        processMultiTenantRequest("credential-options-tenant-id", true, {
           tenantId: "get-token-options-tenant-id"
         }),
         "get-token-options-tenant-id"
@@ -69,17 +64,14 @@ describe("Identity utilities", function() {
 
     it("If we had a tenant and there is no tenant in the options, we pick the tenant we already had", async function() {
       assert.equal(
-        processMultiTenantRequest("credential-options-tenant-id", {
-          allowMultiTenantAuthentication: true
-        }),
+        processMultiTenantRequest("credential-options-tenant-id", true, {}),
         "credential-options-tenant-id"
       );
     });
 
     it("If the tenant received is the same as the tenant we already had, that same tenant should be returned", async function() {
       assert.equal(
-        processMultiTenantRequest("same-tenant", {
-          allowMultiTenantAuthentication: true,
+        processMultiTenantRequest("same-tenant", true, {
           tenantId: "same-tenant"
         }),
         "same-tenant"